Added index.html file
Okay I have kind of changed all the files a little bit, but they work together way nicer now. We have an index.html file which will contain all of the common code between all the pages on the site (e.g. the navbar), so from now on when new pages are created you dont need to include any <html></html> stuff, you just include the stuff that is unique to that page. You also dont need to link to the css as it does that through the index file too now. Each page needs the first line to be {% extends 'index.html' %}, followed by {% load static %}. The content specific to the page is included within {% block content %},{% endblock content %}.