Updated Home page
Updated the home page to display differently depending on if the user is logged in or not. Updated some things on the css file. As a temporary measure, I've changed the text back to black on the log in and sign up pages to make it readable until I get around to restyling them.